fix: Cambiar tipo de datos de agendaId de Long a Integer en OrdenaDTO y ResultadoDiagnostico

+++ EPS/backend/src/main/java/com/eps/apexeps/models/DTOs/OrdenaDTO.java
@@ -1,14 +1,14 @@
 package com.eps.apexeps.models.DTOs;
 
 public class OrdenaDTO {
-    private Long agendaId;
+    private Integer agendaId;
     private String codigoServicio;
 
-    public Long getAgendaId() {
+    public Integer getAgendaId() {
         return agendaId;
     }
 
-    public void setAgendaId(Long agendaId) {
+    public void setAgendaId(Integer agendaId) {
         this.agendaId = agendaId;
     }
